requirements.txt

Flask==3.0.2
python-telegram-bot==20.8
requests==2.31.0
anthropic==0.18.0
googletrans==3.0.0
dateparser==1.2.0
gunicorn==21.2.0
